3.9.241

+++ Managed Source/ArsenalImageMounterAnalyze/Module1.cs
@@ -104,6 +104,6 @@ public static class Module1
     {
         _ = NativeFileIO.PhDiskMntCtl.SendSrbIoControl(SafeFileHandle, NativeFileIO.PhDiskMntCtl.SMP_IMSCSI_QUERY_VERSION, 0U, null, out var ReturnCode);
 
-        return (int)ReturnCode;
+        return ReturnCode;
     }
 }
